test-ANE2/requirements.txt
bidict==0.23.1
certifi==2025.11.12
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
numpy==2.3.5
pyhackrf2==1.0.3
python-crontab==3.3.0
python-engineio==4.12.3
python-socketio==5.15.0
requests==2.32.5
scipy==1.16.3
simple-websocket==1.1.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.3.2
bidict==0.23.1
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-SocketIO==5.5.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-engineio==4.12.3
python-socketio==5.15.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.3.2
